+++ lib/src/modules/instabug.dart
@@ -414,7 +414,7 @@ class Instabug {
   }
 
   /// This API has to be call when using custom app rating prompt
-  static Future<void> willRedirectToAppStore() async {
-    return _host.willRedirectToAppStore();
+  static Future<void> willRedirectToStore() async {
+    return _host.willRedirectToStore();
   }
 }