+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## [Unreleased](https://github.com/Instabug/Instabug-Flutter/compare/v12.7.0...dev) 
+
+### Changed
+
+- Bump Instabug Android SDK to v12.9.0 ([#450](https://github.com/Instabug/Instabug-Flutter/pull/450)). [See release notes](https://github.com/Instabug/Instabug-Android/releases/tag/v12.9.0).
+
+
 ## [12.7.0](https://github.com/Instabug/Instabug-Flutter/compare/v12.5.0...v12.7.0) (February 15, 2024)
 
 ### Added