+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 ### Added
 
 - Support user identification using ID ([#435](https://github.com/Instabug/Instabug-Flutter/pull/435)).
+- add grouping key , error level & user attributes to `CrashReporting.reportHandledCrash`([#455](https://github.com/Instabug/Instabug-Flutter/pull/455)).
+
 
 ### Changed