+++ CHANGELOG.md
@@ -5,7 +5,7 @@
 ### Added
 
 - Adds custom app rating api ([#453](https://github.com/Instabug/Instabug-Flutter/pull/453))
-
+- Add `SessionReplay.getSessionReplayLink` API which retrieves the current session's replay link ([#445](hhttps://github.com/Instabug/Instabug-Flutter/pull/445)).
 
 ## [12.7.0](https://github.com/Instabug/Instabug-Flutter/compare/v12.5.0...v12.7.0) (February 15, 2024)