+++ CHANGELOG.md
@@ -9,7 +9,7 @@
 
 ### Changed
 - Bump Instabug Android SDK to v13.0.0 ([#455](https://github.com/Instabug/Instabug-Flutter/pull/455)). See release notes for [v13.0.0](https://github.com/Instabug/Instabug-Android/releases/tag/v13.0.0).
-
+- Bump Instabug iOS SDK to v13.0.0 ([#446](https://github.com/Instabug/Instabug-Flutter/pull/446)). [See release notes](https://github.com/Instabug/Instabug-iOS/releases/tag/13.0.0).
 
 ## [12.7.0](https://github.com/Instabug/Instabug-Flutter/compare/v12.5.0...v12.7.0) (February 15, 2024)